Fix described class, add minor fixes for specs

+++ backend/app/finders/spree/admin/reports/orders/total_sales.rb
@@ -16,7 +16,7 @@ module Spree
             orders = by_completed_at_min(orders)
             orders = by_completed_at_max(orders)
             orders = grouped_by(orders)
-            values = orders.map { |day, results| [day, results.sum(&:total).to_f] }.sort_by { |day, _| day }.to_h
+            values = orders.map { |day, results| [day, results.sum(&:total).to_i] }.sort_by { |day, _| day }.to_h
 
             labels.map { |label| [label, values[label] || 0] }
           end
